chore: 기존 presentation 레이어 model 을 domain 레이어로 이전
