+++ app/src/main/java/com/yessorae/imagefactory/mapper/UpscaleResultModelMapper.kt
@@ -2,7 +2,7 @@ package com.yessorae.imagefactory.mapper
 
 import com.yessorae.common.replaceDomain
 import com.yessorae.data.remote.stablediffusion.model.response.UpscaleDto
-import com.yessorae.imagefactory.ui.model.UpscaleResultModel
+import com.yessorae.domain.model.UpscaleResultModel
 import javax.inject.Inject
 
 class UpscaleResultModelMapper @Inject constructor() {
