+++ domain/src/main/java/com/yessorae/domain/model/TxtToImgHistory.kt
@@ -1,4 +1,4 @@
-package com.yessorae.imagefactory.ui.model
+package com.yessorae.domain.model
 
 import java.time.LocalDateTime
