+++ domain/src/main/java/com/yessorae/domain/model/TxtToImgRequest.kt
@@ -1,22 +1,23 @@
-package com.yessorae.imagefactory.ui.model
+package com.yessorae.domain.model
 
-data class TxtToImgResultMetaData(
+data class TxtToImgRequest(
     val prompt: String,
     val modelId: String,
     val negativePrompt: String,
-    val w: Int,
-    val h: Int,
+    val width: Int,
+    val height: Int,
     val guidanceScale: Double,
     val seed: Long?,
     val steps: Int,
     val nSamples: Int,
-    val fullUrl: String,
     val upscale: String,
     val multiLingual: String,
     val panorama: String,
     val selfAttention: String,
     val embeddings: String?,
     val lora: String?,
-    val outdir: String?,
-    val filePrefix: String?
+    val loraStrength: String?,
+    val scheduler: String,
+    val safetyChecker: String,
+    val enhancePrompt: String
 )
