+++ domain/src/main/java/com/yessorae/domain/model/UpscaleResultModel.kt
@@ -1,4 +1,4 @@
-package com.yessorae.imagefactory.ui.model
+package com.yessorae.domain.model
 
 data class UpscaleResultModel(
     val id: Int,
