+++ domain/src/main/java/com/yessorae/domain/util/ImageFactoryException.kt
@@ -1,4 +1,4 @@
-package com.yessorae.data.util
+package com.yessorae.domain.util
 
 sealed class ImageFactoryException(override val message: String) : Exception(message) {
     class StableDiffusionApiException(message: String) : ImageFactoryException(message)
